Split landfill_gas.rs (834 lines) into subdirectory modules
Split the monolithic landfill_gas.rs into focused submodules:
- constants.rs (29 lines): LFG generation/collection constants
- state.rs (57 lines): LandfillGasState resource and Default impl
- calculations.rs (76 lines): gas generation, electricity, methane, risk helpers
- systems.rs (86 lines): update_landfill_gas ECS system
- tests_calculations.rs (286 lines): constant and calculation unit tests
- tests_state.rs (265 lines): state defaults, serialization, integration tests
- mod.rs (50 lines): module docs, re-exports, LandfillGasPlugin

All pub items are re-exported from mod.rs so no external imports change.

Part of #1389
